refactor: extract FileList and FileViewer components from index.tsx
