+++ projects/file-server/src/components/FileViewer.tsx
@@ -1,0 +1,9 @@
+import type { FC } from "hono/jsx"
+
+interface FileViewerProps {
+  content: string
+}
+
+export const FileViewer: FC<FileViewerProps> = ({ content }) => {
+  return <pre>{content}</pre>
+}
