+++ projects/file-server/src/index.tsx
@@ -12,8 +12,9 @@ import {
   mkdirHandler,
   uploadFileHandler,
 } from "./api/handlers"
-import { isInvalidPath, sortFiles } from "./utils/fileUtils"
-import { formatFileSize, formatTimestamp } from "./utils/formatters"
+import { FileList } from "./components/FileList"
+import { FileViewer } from "./components/FileViewer"
+import { isInvalidPath } from "./utils/fileUtils"
 
 const app = new Hono<{
   Bindings: {
@@ -147,114 +148,7 @@ app.get("/", async (c) => {
     }
   }
 
-  const sortedFiles = sortFiles(files)
-
-  return c.render(
-    <div>
-      {/* パンくずリストの追加 */}
-      <nav style={{ marginBottom: "1em" }}>
-        {(() => {
-          // requestPathを"/"で分割し、各階層のリンクを生成
-          const parts = requestPath.split("/").filter(Boolean)
-          const crumbs = []
-          let acc = ""
-          // ルート
-          crumbs.push(
-            <span key="root">
-              <a href="/">root</a>
-              {parts.length > 0 ? " / " : ""}
-            </span>,
-          )
-          parts.forEach((part, idx) => {
-            acc += (acc ? "/" : "") + part
-            const isLast = idx === parts.length - 1
-            crumbs.push(
-              <span key={acc}>
-                <a href={`/?path=${encodeURIComponent(acc)}`}>{part}</a>
-                {!isLast ? " / " : ""}
-              </span>,
-            )
-          })
-          return crumbs
-        })()}
-      </nav>
-      <hr />
-      <ul>
-        {sortedFiles.map((file) => (
-          <li
-            key={file.name}
-            style={{
-              display: "flex",
-              justifyContent: "space-between",
-            }}
-          >
-            <a
-              href={`/?path=${encodeURIComponent(
-                path.join(requestPath, file.name),
-              )}`}
-            >
-              {file.name}
-              {file.type === "dir" ? "/" : ""}
-            </a>
-            <div style={{ display: "flex", gap: "1em" }}>
-              {/* ファイルサイズ表示（ファイルのみ） */}
-              {file.type === "file" && (
-                <div
-                  style={{
-                    width: "120px",
-                    textAlign: "right",
-                    margin: "0 1em",
-                  }}
-                >
-                  {formatFileSize(file.size || 0)}
-                </div>
-              )}
-              {/* タイムスタンプ表示 */}
-              <div style={{ width: "180px", textAlign: "right" }}>
-                {file.mtime && formatTimestamp(new Date(file.mtime))}
-              </div>
-              <div
-                style={{ width: "80px", display: "flex", alignItems: "center" }}
-              >
-                <form method="post" action="/api/delete">
-                  <input
-                    type="hidden"
-                    name="path"
-                    value={path.join(requestPath, file.name)}
-                  />
-                  <button type="submit">Delete</button>
-                </form>
-              </div>
-            </div>
-          </li>
-        ))}
-      </ul>
-      <form action="/api/mkdir" method="post" style={{ marginBottom: "1em" }}>
-        <input
-          type="hidden"
-          name="path"
-          value={
-            requestPath
-              ? requestPath + (requestPath.endsWith("/") ? "" : "/")
-              : ""
-          }
-        />
-        <input
-          type="text"
-          name="folder"
-          placeholder="New folder name"
-          required
-          style={{ marginRight: "0.5em" }}
-        />
-        <button type="submit">Create Folder</button>
-      </form>
-      <form action="/api/upload" method="post" enctype="multipart/form-data">
-        <input type="hidden" name="path" value={requestPath} />
-        <input type="file" name="file" required />
-        <button type="submit">Upload</button>
-      </form>
-    </div>,
-  )
+  return c.render(<FileList files={files} requestPath={requestPath} />)
 })
 
 app.get("/file", async (c) => {
@@ -307,7 +201,7 @@ app.get("/file", async (c) => {
 
   if (isText) {
     const content = await readFile(resolvedFile, "utf-8")
-    return c.render(<pre>{content}</pre>)
+    return c.render(<FileViewer content={content} />)
   } else {
     const contentBuffer = await readFile(resolvedFile)
     const isImageOrVideoOrPdf =
